cloud_functions/serp-scraper-function/requirements.txt
functions-framework==3.*
flask==2.*
requests==2.*
google-cloud-bigquery==3.*